Make an abstraction of all agents and move Grakn implementation into a separate package
