+++ grakn/EmploymentAgent.java
@@ -1,6 +1,5 @@
-package grakn.simulation.agents;
+package grakn.simulation.grakn;
 
-import grakn.simulation.agents.common.CityAgent;
 import grakn.simulation.common.ExecutorUtils;
 import graql.lang.Graql;
 import graql.lang.query.GraqlGet;
@@ -9,50 +8,26 @@ import graql.lang.query.GraqlInsert;
 import java.time.LocalDateTime;
 import java.util.List;
 
-import static grakn.simulation.agents.RelocationAgent.cityResidentsQuery;
-import static grakn.simulation.common.Allocation.allocate;
-
-public class EmploymentAgent extends CityAgent {
-
-    private static final double MIN_ANNUAL_WAGE = 18000.00;
-    private static final double MAX_ANNUAL_WAGE = 80000.00;
-    private static final double MIN_CONTRACTED_HOURS = 30.0;
-    private static final double MAX_CONTRACTED_HOURS = 70.0;
-    private static final int MIN_CONTRACT_CHARACTER_LENGTH = 200;
-    private static final int MAX_CONTRACT_CHARACTER_LENGTH = 600;
-
-    private LocalDateTime employmentDate;
+public class EmploymentAgent extends grakn.simulation.agents.interaction.EmploymentAgent {
 
     @Override
-    public void iterate() {
-        employmentDate = today().minusYears(2);
-
-        List<String> employeeEmails;
-        List<Long> companyNumbers;
-        employeeEmails = getEmployeeEmails(employmentDate);
-        companyNumbers = getCompanyNumbers();
-        tx().commit();
-        closeTx();
-        // A second transaction is being used to circumvent graknlabs/grakn issue #5585
-        allocate(employeeEmails, companyNumbers, this::insertEmployment);
-        tx().commit();
-    }
-
-    private List<Long> getCompanyNumbers() {
+    protected List<Long> getCompanyNumbers() {
         GraqlGet companyNumbersQuery = CompanyAgent.getCompanyNumbersInCountryQuery(city().country());
         log().query("getEmployeeEmails", companyNumbersQuery);
         int numCompanies = world().getScaleFactor();
         return ExecutorUtils.getOrderedAttribute(tx().forGrakn(), companyNumbersQuery, "company-number", numCompanies);
     }
 
-    private List<String> getEmployeeEmails(LocalDateTime earliestDate) {
+    @Override
+    protected List<String> getEmployeeEmails(LocalDateTime earliestDate) {
         GraqlGet getEmployeeEmailsQuery = cityResidentsQuery(city(), earliestDate);
         log().query("getEmployeeEmails", getEmployeeEmailsQuery);
         int numEmployments = world().getScaleFactor();
         return ExecutorUtils.getOrderedAttribute(tx().forGrakn(), getEmployeeEmailsQuery, "email", numEmployments);
     }
 
-    private void insertEmployment(String employeeEmail, Long companyNumber){
+    @Override
+    protected void insertEmployment(String employeeEmail, Long companyNumber){
         GraqlInsert insertEmploymentQuery = Graql.match(
                 Graql.var("city")
                         .isa("city")
