+++ grakn/ParentshipAgent.java
@@ -1,7 +1,6 @@
-package grakn.simulation.agents;
+package grakn.simulation.grakn;
 
 import grakn.client.answer.ConceptMap;
-import grakn.simulation.agents.common.CityAgent;
 import grakn.simulation.common.Allocation;
 import graql.lang.Graql;
 import graql.lang.query.GraqlGet;
@@ -19,37 +18,10 @@ import java.util.Map;
 import static grakn.simulation.common.ExecutorUtils.getOrderedAttribute;
 import static java.util.stream.Collectors.toList;
 
-public class ParentshipAgent extends CityAgent {
+public class ParentshipAgent extends grakn.simulation.agents.interaction.ParentshipAgent {
 
     @Override
-    public void iterate() {
-        // Query for married couples in the city who are not already in a parentship relation together
-
-        List<String> childrenEmails = getChildrenEmailsBorn(today());
-
-        List<HashMap<String, String>> marriageEmails = getMarriageEmails();
-
-        if (marriageEmails.size() > 0 && childrenEmails.size() > 0) {
-            LinkedHashMap<Integer, List<Integer>> childrenPerMarriage = Allocation.allocateEvenlyToMap(childrenEmails.size(), marriageEmails.size());
-
-            for (Map.Entry<Integer, List<Integer>> childrenForMarriage : childrenPerMarriage.entrySet()) {
-                Integer marriageIndex = childrenForMarriage.getKey();
-                List<Integer> children = childrenForMarriage.getValue();
-
-                HashMap<String, String> marriage = marriageEmails.get(marriageIndex);
-
-                List<String> childEmails = new ArrayList<>();
-                for (Integer childIndex : children) {
-                    childEmails.add(childrenEmails.get(childIndex));
-                }
-
-                insertParentShip(marriage, childEmails);
-            }
-            tx().commit();
-        }
-    }
-
-    private List<HashMap<String, String>> getMarriageEmails() {
+    protected List<HashMap<String, String>> getMarriageEmails() {
         GraqlGet.Sorted marriageQuery = Graql.match(
                 Graql.var("city").isa("city")
                         .has("location-name", city().name()),
@@ -83,7 +55,8 @@ public class ParentshipAgent extends CityAgent {
                 .collect(toList());
     }
 
-    private List<String> getChildrenEmailsBorn(LocalDateTime dateToday) {
+    @Override
+    protected List<String> getChildrenEmailsBorn(LocalDateTime dateToday) {
 
         GraqlGet.Unfiltered childrenQuery = Graql.match(
                 Graql.var("c").isa("city")
@@ -100,7 +73,8 @@ public class ParentshipAgent extends CityAgent {
         return getOrderedAttribute(tx().forGrakn(), childrenQuery, "email");
     }
 
-    private void insertParentShip(HashMap<String, String> marriage, List<String> childEmails) {
+    @Override
+    protected void insertParentShip(HashMap<String, String> marriage, List<String> childEmails) {
         ArrayList<Statement> matchStatements = new ArrayList<>(Arrays.asList(
                 Graql.var("mother").isa("person").has("email", marriage.get("wife-email")),
                 Graql.var("father").isa("person").has("email", marriage.get("husband-email"))
