+++ grakn/PersonBirthAgent.java
@@ -1,45 +1,12 @@
-package grakn.simulation.agents;
+package grakn.simulation.grakn;
 
-import grakn.simulation.agents.common.CityAgent;
 import graql.lang.Graql;
 import graql.lang.query.GraqlInsert;
 
-public class PersonBirthAgent extends CityAgent {
+public class PersonBirthAgent extends grakn.simulation.agents.interaction.PersonBirthAgent {
 
     @Override
-    public void iterate() {
-        // Find bachelors and bachelorettes who are considered adults and who are not in a marriage and pair them off randomly
-        int numBirths = world().getScaleFactor();
-        for (int i = 0; i < numBirths; i++) {
-            String gender;
-            String forename;
-            String surname = pickOne(world().getSurnames());
-
-            boolean genderBool = random().nextBoolean();
-            if (genderBool) {
-                gender = "male";
-                forename = pickOne(world().getMaleForenames());
-            } else {
-                gender = "female";
-                forename = pickOne(world().getFemaleForenames());
-            }
-
-            // Email is used as a key and needs to be unique, which requires a lot of information
-            String email = forename + "."
-                    + surname + "_"
-                    + today().toString() + "_"
-                    + i + "_"
-                    + simulationStep() + "_"
-                    + city() + "_"
-                    + city().country() + "_"
-                    + city().country().continent()
-                    + "@gmail.com";
-            insertPerson(email, gender, forename, surname);
-        }
-        tx().commit();
-    }
-
-    public void insertPerson(String email, String gender, String forename, String surname) {
+    protected void insertPerson(String email, String gender, String forename, String surname) {
 
 //        Result result = tx().run("MATCH (c:City), c.location-name = $location-name" +
 //                "CREATE (p:Person {" +
@@ -80,6 +47,4 @@ public class PersonBirthAgent extends CityAgent {
         log().query("insertPerson", query);
         tx().forGrakn().execute(query);
     }
-
-
 }
